% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{as.data.frame.coleoGetResp}
\alias{as.data.frame.coleoGetResp}
\title{Transformer la liste de réponses en data.frame}
\usage{
\method{as.data.frame}{coleoGetResp}(responses = responses)
}
\arguments{
\item{responses}{liste de réponse des fonctions de GET/POST}
}
\description{
Transformer la liste de réponses en data.frame
}
