% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_color_palettes.R
\name{bq_pal_qual}
\alias{bq_pal_qual}
\title{Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec qualitative}
\usage{
bq_pal_qual(size = 5)
}
\arguments{
\item{size}{nombre de couleurs à retourner, par défaut 5}
}
\value{
vecteur de couleurs hexadécimales
}
\description{
Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec qualitative
}
\examples{
library(scales); show_col(bq_pal_qual(8))
}
