% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_color_palettes.R
\name{bq_pal_quant_bi}
\alias{bq_pal_quant_bi}
\title{Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec quantitative}
\usage{
bq_pal_quant_bi(color1 = "vert", color2 = "bleu", ton = "base", size = 10)
}
\arguments{
\item{color1}{couleur de base 1. "vert", "bleu", "orange", "gris", "beige" par défaut "vert"}

\item{color2}{couleur de base 2, par défaut "bleu"}

\item{ton}{ton de couleur. Base, fonce ou clair. Par défaut "base"}

\item{size}{nombre de couleurs à retourner, par défaut 10}
}
\value{
vecteur de couleurs hexadécimales
}
\description{
Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec quantitative
}
\examples{
library(scales); show_col(bq_pal_quant_bi("vert", "bleu", "base", 10))
library(scales); show_col(bq_pal_quant_bi("orange", "bleu", "fonce", 10))
}
