% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_color_palettes.R
\name{bq_pal_quant_mono}
\alias{bq_pal_quant_mono}
\title{Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec quantitative}
\usage{
bq_pal_quant_mono(color = "orange", size = 10)
}
\arguments{
\item{color}{couleur de base, par défaut "orange"}

\item{size}{nombre de couleurs à retourner, par défaut 10}
}
\value{
vecteur de couleurs hexadécimales
}
\description{
Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec quantitative
}
\examples{
library(scales); show_col(bq_pal_quant_mono("orange", 10))
}
