% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_color_palettes.R
\name{bq_pal_quant_tri}
\alias{bq_pal_quant_tri}
\title{Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec quantitative}
\usage{
bq_pal_quant_tri(
  color1 = "orange",
  color2 = "bleu",
  color3 = "vert",
  ton = "base",
  size = 10
)
}
\arguments{
\item{color1}{couleur de base 1. "orange", "bleu", "vert" par défaut "orange"}

\item{color2}{couleur de base 2, par défaut "bleu"}

\item{color3}{couleur de base 3, par défaut "vert"}

\item{ton}{ton de couleur. Base, fonce ou clair. Par défaut "base"}

\item{size}{nombre de couleurs à retourner, par défaut 10}
}
\value{
vecteur de couleurs hexadécimales
}
\description{
Retourne un vecteur de valeurs hex pour des couleurs Biodiversité Québec quantitative
}
\examples{
library(scales); show_col(bq_pal_quant_tri("orange", "bleu", "vert", "base", 10))
library(scales); show_col(bq_pal_quant_tri("orange", "bleu", "vert", "fonce", 10))
}
