% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cells}
\alias{cells}
\title{Plant and pollinator example network}
\format{Une liste équivalente au format JSON pour des injections dans l'API de Coléo}
\usage{
data(cells)
}
\description{
Cells est un jeu de données contenant les 9 000 cellules géoréférencées du plan d'échantillonage établie en 2016
}
\details{
\itemize{
\item cell_code: Code d'identification unique de la cellule du plan d'échantillongae (TODO: Ref Samson 2016)
\item name: Nom unique de la cellule (si la cellule a fait l'objet de la mise en place de site).
\item shp: Fichier de forme de la cellule
}
}
\author{
MFFP 2017
}
\keyword{datasets}
