% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{cells_sf}
\alias{cells_sf}
\title{Fonction retournant les cellules dans un objet de classe spatiale (sf)}
\usage{
cells_sf(cell_code = NULL, name = NULL, ...)
}
\arguments{
\item{cell_code}{\code{character} vecteur contenant les identifiants uniques que l'on désire obtenir. Si cell_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.}

\item{name}{\code{character} vecteur contenant les noms de la cellules que l'on veut obtenir.}

\item{...}{arguments de la fonction \code{httr::GET()} et \code{rcoleo::get_cells()}}
}
\description{
Fonction retournant les cellules dans un objet de classe spatiale (sf)
}
\seealso{
\code{\link[=get_cells]{get_cells()}}
}
