% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{cl_to_sf.coleoGet}
\alias{cl_to_sf.coleoGet}
\title{Fonction générique pour convertir les sorties de l'API en objet de classe spatiale (sf)}
\usage{
cl_to_sf.coleoGet(responses = responses)
}
\arguments{
\item{responses}{\code{list} objet retourné par les fonctions GET du package coléo (ex. get_sites()), voir @details}
}
\value{
Retourne un objet de type \code{list} contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est \code{getSuccess} ou \code{getError}. Une réponse de classe \code{getSuccess} est une liste à deux niveaux composé du contenu (\code{body}), et la réponse \link[httr:response]{httr::response}. Une réponse de classe \code{getError} dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.
}
\description{
Fonction générique pour convertir les sorties de l'API en objet de classe spatiale (sf)
}
\details{
L'objet d'entrée doit contenir les colonnes \code{geom.type} & \code{geom.coordinates}
}
\examples{
resp <- get_sites()
plot(cl_to_sf(resp))
}
\seealso{
\code{\link[=cells_sf]{cells_sf()}}, \code{\link[=sites_cf]{sites_cf()}}
}
