% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_convert.R
\name{cl_to_sf}
\alias{cl_to_sf}
\alias{cl_to_sf.coleoGetResp}
\alias{cl_to_sf.data.frame}
\title{Fonction générique pour convertir les sorties de l'API en objet de classe spatiale (sf)}
\usage{
cl_to_sf(data)

\method{cl_to_sf}{coleoGetResp}(data)

\method{cl_to_sf}{data.frame}(data)
}
\arguments{
\item{data}{\code{list} de la classe \code{coleoGetResp} retourné par les fonctions GET du package coléo (ex. \code{\link[=get_sites]{get_sites()}}) ou \code{data.frame} (\link{as.data.frame}) , voir @details}
}
\value{
Retourne un objet de la classe \code{sf::st_sf}
}
\description{
Fonction générique pour convertir les sorties de l'API en objet de classe spatiale (sf)
}
\details{
L'objet d'entrée doit contenir les colonnes \code{geom.type} & \code{geom.coordinates}
}
\examples{
\dontrun{
plot(cl_to_sf(get_sites()))
} 
}
\seealso{
\code{\link[=sf_cells]{sf_cells()}}, \code{\link[=sf_sites]{sf_sites()}}
}
