% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo2_request.R
\name{coleo2_extract_id}
\alias{coleo2_extract_id}
\title{Extract an id from an API response
Extraction de l'id d'une réponse de l'API de coleo}
\usage{
coleo2_extract_id(answer_resp)
}
\arguments{
\item{answer_resp}{Réponse httr2 de l'API de coleo.}
}
\value{
Le id sous format d'integer ou \code{NA_integer_}
}
\description{
Fonction de commodité pour processer les réponses de l'API de coleo.
Fonctionnelle pour les nouvelles entrées (réponses de requêtes POST) et
pour celles qui existent déjà (requêtes GET).
}
