% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_api_utils.R
\name{coleo_begin_req}
\alias{coleo_begin_req}
\title{Pépare une requête générique à l'API de coleo}
\usage{
coleo_begin_req(schema)
}
\arguments{
\item{schema}{Schéma sur lequel faire la requête}
}
\value{
Un objet httr2 request
}
\description{
Pépare une requête générique à l'API de coleo
}
