% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_cell.R
\name{coleo_cell_geom_fmt}
\alias{coleo_cell_geom_fmt}
\title{Converti les cellules du format sf (objet sf) en un un objet list.}
\usage{
coleo_cell_geom_fmt(cell_geom)
}
\arguments{
\item{cell_geom}{La géométrie d'une cellule en tant qu'objet sf de type 'polygon'.}
}
\value{
Une liste imbriquée contenant les coordonnées du polygone.
}
\description{
Les cellules doivent être injectées en tant que listes imbriquées. La
lecture des sites en R se fait généralement en utilisant sf::read_sf().
Cette fonction converti l'objet sf en une liste imbriquée via geojsonio.
}
