% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLEO_comp.R
\name{COLEO_comp}
\alias{COLEO_comp}
\title{Compare deux vecteurs et donne des détails si ils ne sont pas égaux}
\usage{
COLEO_comp(x, y)
}
\value{
un message
}
\description{
Compare deux vecteurs et donne des détails si ils ne sont pas égaux
}
\examples{
\dontrun{
# Vecteurs égaux
i <- 1:5
j <- 1:10

COLEO_comp(i, j)

# Vecteurs inégaux
i <- 1:13
j <- 1:10

COLEO_comp(i, j)
}
}
