% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_execute_injection}
\alias{coleo_execute_injection}
\title{Inject data into coleo}
\usage{
coleo_execute_injection(df)
}
\arguments{
\item{df}{a dataframe with one (and only one) column containing http requests
to inject data}
}
\value{
the same dataframe, with three new columns: the http response, the
error message (if any) and a column \code{success} which is TRUE if that
row was successfully injected
}
\description{
Takes a dataframe with a column of httr2 POST requests and actually performs
them
}
