% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_extract_id}
\alias{coleo_extract_id}
\title{Extract an id from an API response}
\usage{
coleo_extract_id(answer_resp)
}
\arguments{
\item{answer_resp}{an HTTP response from the coleo API}
}
\value{
the ID as an integer or \code{NA_integer_}
}
\description{
this is a convenience function for processing the response from the COLEOs
api. It should work for both new records (the response from POST requests)
and existing ones (GET requests).
}
