% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_format_utils.R
\name{coleo_format_extra_col}
\alias{coleo_format_extra_col}
\title{Preparer la colonne extra pour injection}
\usage{
coleo_format_extra_col(df, db_table)
}
\arguments{
\item{df}{le dataframe à preparer}

\item{db_table}{la table ciblée dans la base de données}
}
\value{
un tbl
}
\description{
Preparer la colonne extra pour injection
}
