% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get.R
\name{coleo_get_attributes_table}
\alias{coleo_get_attributes_table}
\title{Retourne la table attributes ou les valeurs d'un champ si spécifié}
\usage{
coleo_get_attributes_table(column = NULL)
}
\arguments{
\item{column}{Un champ de la table attributes}
}
\value{
un data frame de la table attributes ou des valeurs contenues dans un champ
}
\description{
Retourne la table attributes ou les valeurs d'un champ si spécifié
}
