% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_get_cell_id}
\alias{coleo_get_cell_id}
\title{Download the cell id}
\usage{
coleo_get_cell_id(cell_code)
}
\arguments{
\item{site_code}{code pour la cellule, du format "123_92"}
}
\value{
code for cell (integer) or integer NA value
}
\description{
Download the cell id
}
