% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_name_table.R
\name{coleo_get_name_table}
\alias{coleo_get_name_table}
\title{Table contenant pour chaque table de la base de données les noms de colonnes, les champs correspondants dans la base de données, les champs obligatoires, la classe des données par colonne et les valeurs possibles par colonnes}
\usage{
coleo_get_name_table()
}
\value{

}
\description{
Table contenant pour chaque table de la base de données les noms de colonnes, les champs correspondants dans la base de données, les champs obligatoires, la classe des données par colonne et les valeurs possibles par colonnes
}
