% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_get_site_id}
\alias{coleo_get_site_id}
\title{Download the site id}
\usage{
coleo_get_site_id(site_code)
}
\arguments{
\item{site_code}{code pour le site, du format "123_92_F01"}
}
\value{
code for site (integer) or integer NA value
}
\description{
Download the site id
}
