% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_tsn.R
\name{coleo_get_tsn}
\alias{coleo_get_tsn}
\title{Fonction d'aide à l'obtention des tsn}
\usage{
coleo_get_tsn(taxa, tsn_rows = NA)
}
\arguments{
\item{taxa}{nom scientifique des espèces}

\item{tsn_rows}{numérique, chiffre de 1 à infini. La valeur par défault NA spécifie que toutes les identifications}
}
\value{

}
\description{
Fonction d'aide à l'obtention des tsn
}
