% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject}
\alias{coleo_inject}
\title{Inject data into coleo}
\usage{
coleo_inject(df)
}
\arguments{
\item{df}{a valid dataframe}
}
\value{
A message specifying the dataframe lines that were injected and that
failed to be injected, and the resulting dataframes from
coleo_injection_perform()
}
\description{
Takes a valid dataframe and performs autonomously the injection of all tables
}
