% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_adne_landmarks}
\alias{coleo_inject_adne_landmarks}
\title{Injection des repères d'une campagne ADNe dans la table
landmarks de coleo.}
\usage{
coleo_inject_adne_landmarks(df_id, campaign_type, schema = "public")
}
\arguments{
\item{df_id}{Un data.frame contenant une colonne campaign_id.}

\item{campaign_type}{Type de la campagne. Doit être "ADNe".}

\item{schema}{Schéma sur lequel faire la requête. Par défaut, le schéma
public est utilisé.}
}
\value{
Une data.frame avec une colonne landmark_X_id et une colonne pour
les landmark_X_error.
}
\description{
L'injection des repèes ADNe est différente des autres de par la structure du jeu de données (tel que formaté par le template). Les observation à l'échelle du lac n'ont pas de repères alors que celles à l'échelle de la station ont des repères.
}
\details{
Accepte un data.frame validè par \code{\link[rcoleo]{coleo_validate}} et performe
l'injection de la table landmarks.

La fonction est utilisée par \code{\link[rcoleo]{coleo_inject}}.
}
