% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_cells}
\alias{coleo_inject_cells}
\title{Injection des cellules dans Coleo.}
\usage{
coleo_inject_cells(df)
}
\arguments{
\item{df}{Un data.frame contenant une colonne \code{geom} de type polygon, cell_code et name.}
}
\description{
Accepte un data.frame validé par \code{coleo_validate} et performe
l'injection dans la table cells.
}
\details{
Cette fonction fait appel aux fonctions de préparation, d'exécution et de
finalisation de l'injection : \code{coleo_inject_general_df},
\code{coleo_injection_execute} et \code{coleo_injection_final}.
}
