% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_general}
\alias{coleo_inject_general}
\title{General injection into the Coleo database}
\usage{
coleo_inject_general(..., endpoint)
}
\arguments{
\item{...}{the data to be injected. May contain NA or NULL values; these will
be dropped before injection.}

\item{endpoint}{The name of the endpoint you are injecting.}
}
\value{
an httr2 request, ready to be performed
}
\description{
General injection into the Coleo database
}
\examples{
coleo_inject_general("cells", cell_code = "444_33", name = "Middle Earth", geom = "")
}
