% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_general_df}
\alias{coleo_inject_general_df}
\title{Inject one dataframe row}
\usage{
coleo_inject_general_df(df_one_row, endpoint)
}
\arguments{
\item{df_one_row}{one row of the dataframe. This can be passed in with
\code{\link[dplyr]{cur_data_all}}}

\item{endpoint}{endpoint for coleo DB. Should name the table where you want
to inject this.}
}
\value{
A single HTTP POST request. this is NOT run. You have to run it
manually. First, you should inspect the contents with
\code{\link[httr2]{req_dry_run}}
}
\description{
Inject one dataframe row
}
