% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_mam_landmarks}
\alias{coleo_inject_mam_landmarks}
\title{Inject landmarks table of mammifère campaigns into coleo}
\usage{
coleo_inject_mam_landmarks(df_id)
}
\arguments{
\item{df_id}{a dataframe with campaign_id column}
}
\value{
a dataframe with landmarks ids
}
\description{
Takes a valid dataframe and performs autonomously the injection of landmarks table
}
