% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_media}
\alias{coleo_inject_media}
\title{Injection of media files into the Coleo server}
\usage{
coleo_inject_media(df_id, failures, server_dir = "observation", file_dir)
}
\arguments{
\item{failures}{Boolean. Failures within the injection process.}

\item{dir}{The name of the directory you are injecting (campaign, observation, site).}

\item{id}{The id generated by coleo for the table row to link to the file. Can be the observation_id, the campaign_id, or the site_id.}

\item{file_path}{The local path to the media file to inject.}
}
\value{
an httr2 request, ready to be performed.
}
\description{
Injection of media files into the Coleo server
}
