% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_ref_species}
\alias{coleo_inject_ref_species}
\title{Inject taxa_names in ref_species table of coleo}
\usage{
coleo_inject_ref_species(taxa_col)
}
\arguments{
\item{taxa_col}{a dataframe with one column of taxa_names}
}
\description{
Takes a vector of taxa_names and autonomously attempts to
inject them in ref_species table.
}
\details{
This function is silent and won't print any messages
as all taxa_names will be POSTed to the server, but all
taxons already present will return an error message.
}
