% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_vegetation_transect_campaigns}
\alias{coleo_inject_vegetation_transect_campaigns}
\title{Injection des campagnes végétation_transect dans la table campaigns de coleo.}
\usage{
coleo_inject_vegetation_transect_campaigns(df_id, schema = "public")
}
\arguments{
\item{df_id}{Un data.frame contenant les données de végétation_transect à injecter.}

\item{schema}{Schéma sur lequel faire la requête.}
}
\value{
Une data.frame avec une colonne campaign_id et une colonne pour
les campaign_error.
}
\description{
Les campagnes de transects de végétation nécessitent tout d'abord de vérifier si les campagnes existent.
}
\details{
Accepte un data.frame validé par \code{\link[rcoleo]{coleo_validate}} et performe
l'injection de la table campaigns.

La fonction est utilisée par \code{\link[rcoleo]{coleo_inject}}.
}
