% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_injection_final}
\alias{coleo_injection_final}
\title{Finalise l'injection dans Coleo.}
\usage{
coleo_injection_final(df)
}
\arguments{
\item{df}{Une data.frame produit par \code{coleo_injection_execute}}
}
\value{
Un data.frame. Sans les groupes et les données imbriquées.
}
\description{
Cette fonction extrait les nouveaux ID et les messages d'erreur des réponses
des injections de coleo. Les ID et erreurs sont sauvés dans une colonne nommée selon la
convention 'table_champ' dans le jeu de données d'origine.
}
\details{
Son utilisation suit celle de \code{\link[rcoleo]{coleo_injection_execute}}
dans le processus d'injection.

La fonction est utilisée dans \code{\link[rcoleo]{coleo_inject_table}},
\code{\link[rcoleo]{coleo_inject_media}} et
\code{\link[rcoleo]{coleo_inject_mam_landmarks}}
}
