% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_injection_prep}
\alias{coleo_injection_prep}
\title{Prepare dataset for injection}
\usage{
coleo_injection_prep(df, db_table)
}
\arguments{
\item{df}{dataframe containing injection material.}

\item{db_table}{the database table into which the data will be injected}
}
\value{
the same dataframe, but with a new column for injection request. Note
that the dataframe will be \code{rowwise} after this function
}
\description{
Take a dataset and prepare the data for injection into a
}
