% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_make.R
\name{coleo_make_df_column_names}
\alias{coleo_make_df_column_names}
\title{Fonction pour standardiser la production les noms de colonnes des jeux de données à importer}
\usage{
coleo_make_df_column_names(db_table, db_colname)
}
\arguments{
\item{db_table}{une table de la base de données.}

\item{db_colname}{les noms de champs de la table.}
}
\value{

}
\description{
Fonction pour standardiser la production les noms de colonnes des jeux de données à importer
}
