% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_api_utils.R
\name{coleo_media_begin_req}
\alias{coleo_media_begin_req}
\title{Pépare une requête générique à l'API de coleo-media}
\usage{
coleo_media_begin_req(server_dir)
}
\arguments{
\item{server_dir}{Dossier sur le serveur où faire la requête}
}
\value{
Un objet httr2 request
}
\description{
Pépare une requête générique à l'API de coleo-media
}
