% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_api_utils.R
\name{coleo_plumber_update}
\alias{coleo_plumber_update}
\title{Déclencheur de l'API plumber pour mettre à jour les données du portail}
\usage{
coleo_plumber_update()
}
\value{
Void. Ouvre le navigateur à l'endpoint de mise à jour.
}
\description{
Cette fonction déclenche la mise à jour des données du portail.
}
