% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_prep_input_data.R
\name{coleo_prep_input_data}
\alias{coleo_prep_input_data}
\title{Preparer les données pour injection}
\usage{
coleo_prep_input_data(df, db_table, site_id = FALSE, schema = "public")
}
\arguments{
\item{df}{le dataframe à preparer}

\item{db_table}{la table ciblée dans la base de données}

\item{site_id}{si TRUE, ajoute la colonne site_id
(par défaut FALSE)}

\item{schema}{le schéma de la base de données}
}
\value{
un tbl, nested
}
\description{
Est utilisé par \code{\link[rcoleo]{coleo_injection_prep}} pour formater
les données à injecter et les données à transformer (géométries,
colonnes extra). Imbrique et isole des colonnes à injecter, place les autres
dans une colonne 'data'.
}
\details{
Cette colonne fait appel aux fonctions de formatage
\code{\link[rcoleo]{coleo_format_extra_col}} et
\code{\link[rcoleo]{format_spatial}}.
}
