% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_read.R
\name{coleo_read}
\alias{coleo_read}
\title{Importe les données à injecter depuis un template coleo ou un document csv
et formate les colonnes pour l'injection dans la base de données coleo.}
\usage{
coleo_read(filePath)
}
\arguments{
\item{filePath}{Chemin local vers le document de données.}
}
\value{
Un data.frame avec les colonnes formatées pour l'injection à coleo.
}
\description{
Importe les données et formate la classe des colonnes pour injection dans la base de données coleo.
}
