% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_read.R
\name{coleo_read_csv}
\alias{coleo_read_csv}
\title{Importe les données à injecter depuis un document csv.}
\usage{
coleo_read_csv(fileName)
}
\arguments{
\item{fileName}{Chemin local vers le document csv.}
}
\value{
Un data.frame avec les colonnes formatées pour l'injection à coleo.
}
\description{
Importe les données à injecter depuis un document csv.
}
