% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_read.R
\name{coleo_read_data}
\alias{coleo_read_data}
\title{Reads a data file and format columns to coleo requirements}
\usage{
coleo_read_data(filePath)
}
\arguments{
\item{filePath}{Local path to a data file.}
}
\value{
data.frame with formated columns
}
\description{
Reads a data file and format columns to coleo requirements
}
