% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_read.R
\name{coleo_read_shape}
\alias{coleo_read_shape}
\title{Importe les données de cellules à injecter depuis un shapefile.}
\usage{
coleo_read_shape(filePath)
}
\arguments{
\item{filePath}{Chemin local vers le document csv.}
}
\value{
Un data.frame avec les colonnes formatées pour l'injection à coleo.
}
\description{
Importe les données de cellules à injecter depuis un shapefile.
}
