% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_read.R
\name{coleo_read_template}
\alias{coleo_read_template}
\title{Read data directly from a coleo template and format columns to coleo requirements}
\usage{
coleo_read_template(templatePath)
}
\arguments{
\item{templatePath}{Local path to template file.}
}
\value{
a data.frame with formated columns
}
\description{
Read data directly from a coleo template and format columns to coleo requirements
}
