% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_request_by_code}
\alias{coleo_request_by_code}
\title{Request a database record by code}
\usage{
coleo_request_by_code(human_code, table, perform = TRUE)
}
\arguments{
\item{human_code}{human-readable database code}

\item{table}{the database table to look for. Right now this is either "cells" or "sites"}

\item{perform}{Should the request be performed? defaults to TRUE.}
}
\value{
if perform = TRUE, the answer is returned. if perform = FALSE, an httr2 query is returned.
}
\description{
many of the database records have unique human-readable codes. Site codes
look like "CCC_CC_LCC" (C = chiffre, L=lettre) and cell codes look like
"CCC_CC". this function takes one of these "human codes" and returns the
information on that record.
}
