% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_request_data}
\alias{coleo_request_data}
\title{Requête de type 'GET' sur les tables de données par type d'inventaire de coleo}
\usage{
coleo_request_data(survey_type, view = "short", ...)
}
\arguments{
\item{survey_type}{Type d'inventaire. Les types d'inventaires supportés sont 'vegetation', 'acoustique_anoures', 'acoustique_chiropteres', 'acoustique_oiseaux', 'acoustique_orthopteres' et 'adne_corrige', 'benthos', 'decomposition_sol', 'insectes_sol', 'mammiferes', 'odonates', 'papilionides', 'physicochimie_terrain', 'zooplancton'.}

\item{view}{'short' par défaut. Type de vue des données à retourner. 'short' retourne les données de base, 'long' retourne les données dans le format gabarit d'injection (requiert accès prévilégier).}

\item{...}{Paramètres de requête pour la base de données coleo (dans le format 'nom' = 'valeur')}
}
\value{
Tibble contenant les données de l'inventaire.
}
\description{
Cette fonction permet de faire une requête sur les tables de données de l'API de coleo.
}
\examples{
# Requête pour les inventaires de type 'végétation'
coleo_request_data(survey_type = 'vegetation', view = 'short')

}
