% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_resp_df}
\alias{coleo_resp_df}
\title{Returns a request to a tibble}
\usage{
coleo_resp_df(resp, written_req, limit)
}
\arguments{
\item{resp}{httr2 response from coleo.}

\item{written_req}{request passed to coleo's api.}

\item{perform}{Should the request be performed? defaults to TRUE.}
}
\value{
if perform = TRUE, the answer is returned. if perform = FALSE, an httr2 query is returned.
}
\description{
If a response contains multiple pages, the function gets them all
and joins them in a single dataframe.
}
