% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_return.R
\name{coleo_return_campaign_type}
\alias{coleo_return_campaign_type}
\title{Retourne le 'campaign_type' d'un jeu de données pour tous types d'inventaires.}
\usage{
coleo_return_campaign_type(data)
}
\arguments{
\item{data}{un jeu de données}
}
\value{
le 'campaign_type' du jeu de données
}
\description{
Retourne le 'campaign_type' d'un jeu de données pour tous types d'inventaires.
}
