% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_required_tables_functions.R
\name{coleo_return_cols}
\alias{coleo_return_cols}
\title{Trouver les colonnes requises, leur classe et les valeurs admissibles pour un type de campagne donné}
\usage{
coleo_return_cols(campaign_type)
}
\arguments{
\item{campaign_type}{character string. Un type de campagne valide.}
}
\value{

}
\description{
Trouver les colonnes requises, leur classe et les valeurs admissibles pour un type de campagne donné
}
