% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_required_tables_functions.R
\name{coleo_return_df_cols}
\alias{coleo_return_df_cols}
\title{Trouver les colonnes requises, leur classe et les valeurs admissibles pour un type de campagne donné}
\usage{
coleo_return_df_cols(camp_type, required_only = TRUE)
}
\arguments{
\item{camp_type}{un type de campagne valide.}

\item{required_only}{boolean. Si FALSE, retourne toutes les colonnes admissibles pour le type de campagne. Autrement, retourne que les colonnes requises pour une injection valide}
}
\value{

}
\description{
Trouver les colonnes requises, leur classe et les valeurs admissibles pour un type de campagne donné
}
