% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_return_cols.R
\name{coleo_return_required_cols}
\alias{coleo_return_required_cols}
\title{Trouver les colonnes requises, leur classe et les valeurs admissibles pour un type de campagne donné}
\usage{
coleo_return_required_cols(campaign_type)
}
\arguments{
\item{campaign_type}{un type de campagne}
}
\description{
Trouver les colonnes requises, leur classe et les valeurs admissibles pour un type de campagne donné
}
