% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_name_table.R
\name{coleo_return_valid_site_type}
\alias{coleo_return_valid_site_type}
\title{Retourne un vecteur contenant les noms valides de type de site}
\usage{
coleo_return_valid_site_type()
}
\value{

}
\description{
Retourne un vecteur contenant les noms valides de type de site
}
