% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_return.R
\name{coleo_return_valid_site_types}
\alias{coleo_return_valid_site_types}
\title{Retourne un vecteur contenant les noms valides de type de site}
\usage{
coleo_return_valid_site_types()
}
\value{
vecteur de charactères contenant tous les types de sites valides
}
\description{
Retourne un vecteur contenant les noms valides de type de site
}
