% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{coleo_validate}
\alias{coleo_validate}
\title{Title}
\usage{
coleo_validate(data, campaign_type)
}
\arguments{
\item{data}{dataset to validate}

\item{campaign_type}{type of campaign to be injected.}
}
\value{

}
\description{
Title
}
