% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{coleo_validate_coordinates}
\alias{coleo_validate_coordinates}
\title{Valider les Coordonnées}
\usage{
coleo_validate_coordinates(
  data,
  dat_names,
  bbox = c(-79.76, 44.99, -57.1, 62.59)
)
}
\arguments{
\item{data}{Un dataframe contenant les coordonnées à valider.}

\item{dat_names}{Un vecteur de caractères des noms de colonnes dans le dataframe qui contiennent les valeurs de latitude et de longitude.}

\item{bbox}{Un vecteur de 4 éléments représentant la boîte englobante des coordonnées géographiques. Par défaut, la boîte englobante du Québec est \link{-79.76, 44.99, -57.10, 62.59}.}
}
\value{
Cette fonction retourne un message d'avertissement si des coordonnées sont hors des plages valides.
}
\description{
Cette fonction valide les coordonnées dans un jeu de données donné pour s'assurer qu'elles se situent dans la boîte englobante du Québec.
}
\details{
La fonction vérifie que les valeurs de latitude et de longitude se situent dans la boîte englobante du Québec (xmin: -79.76, ymin: 44.99, xmax: -57.10, ymax: 62.59).
}
\examples{
\dontrun{
data <- data.frame(lat = c(45, 50, 60), lon = c(-70, -75, -80))
dat_names <- c("lat", "lon")
bbox <- c(-79.76, 44.99, -57.10, 62.59)
coleo_validate_coordinates(data, dat_names, bbox)
}
}
