% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{coleo_validate_coordinates_projection}
\alias{coleo_validate_coordinates_projection}
\title{Valider les Coordonnées}
\usage{
coleo_validate_coordinates_projection(
  data,
  dat_names,
  bbox = c(-90, 90, -180, 180)
)
}
\arguments{
\item{data}{Un dataframe contenant les coordonnées à valider.}

\item{dat_names}{Un vecteur de caractères des noms de colonnes dans le dataframe qui contiennent les valeurs de latitude et de longitude.}

\item{bbox}{Un vecteur de 4 éléments représentant la boîte englobante des coordonnées géographiques. Par défaut, la boîte englobante est \link{-90, 90, -180, 180}.}
}
\value{
Cette fonction retourne un message d'avertissement si des coordonnées sont hors des plages valides.
}
\description{
Cette fonction valide les coordonnées dans un jeu de données donné pour s'assurer qu'elles se situent dans des plages valides du système de coordonnées géographique EPSG:4326.
}
\details{
La fonction valide que les valeurs de latitude sont dans la plage \link{-90, 90} et que les valeurs de longitude sont dans la plage \link{-180, 180}.
}
\examples{
\dontrun{
data <- data.frame(lat = c(45, 50, 60), lon = c(-70, -75, -80))
dat_names <- c("lat", "lon")
bbox <- c(-90, 90, -180, 180)
coleo_validate_coordinates(data, dat_names)
}
}
