% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{coleo_validate_date_digits}
\alias{coleo_validate_date_digits}
\title{Validation du format des dates}
\usage{
coleo_validate_date_digits(data, dat_names, cols_date)
}
\arguments{
\item{data}{Le dataframe contenant les données à injecter.}

\item{dat_names}{Les noms des colonnes du dataframe.}

\item{cols_date}{Les noms des colonnes contenant des date.}
}
\value{
Le message de validation.
}
\description{
Validation du format des dates
}
