% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{coleo_validate_empty_cols}
\alias{coleo_validate_empty_cols}
\title{Validation des colonnes vides}
\usage{
coleo_validate_empty_cols(data, columns)
}
\arguments{
\item{data}{Le dataframe contenant les données à injecter.}

\item{columns}{Les noms des colonnes à valider.}
}
\value{
Le vecteur booléen représentant si la colonne est vide.
}
\description{
Validation des colonnes vides
}
