% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{diagnostic}
\alias{diagnostic}
\title{Effectuer un diagnotic rapide de l'ensemble des appels sur l'API}
\usage{
diagnostic(responses)
}
\arguments{
\item{responses}{liste de réponse des fonctions de GET/POST}
}
\description{
Effectuer un diagnotic rapide de l'ensemble des appels sur l'API
}
