% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_campaigns.R
\name{download_all_campaigns}
\alias{download_all_campaigns}
\title{Download all the campaigns}
\usage{
download_all_campaigns(campaign_df, camp_col = "campaigns")
}
\arguments{
\item{campaign_df}{A data frame containing campaign information}

\item{camp_col}{The name of the column holding the campaigns. Defaults to
"campaigns" which is what the API returns}
}
\value{
A data.frame just like the \code{campaign_df}, but with a list-column
which contains the response. This column is called \code{campaign_responses}
}
\description{
Download every campaign described in a data.frame of sites
}
\details{
This function is designed to work with the same kind of object returned by
\code{\link[=download_sites_sf]{download_sites_sf()}}. However it does not need to be an sf data.frame. Any
data.frame with a column containing campaign information \emph{should} work.
}
