% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_campaigns.R
\name{download_all_requests}
\alias{download_all_requests}
\title{Download all queries stored in a list}
\usage{
download_all_requests(
  starting_df,
  request_info_col = "campaigns",
  query_info,
  endpoint,
  request_col_name
)
}
\arguments{
\item{starting_df}{A data frame containing info about a table in the API. You
start with this, and you want to add the new information to it}

\item{request_info_col}{The name of the column holding the campaigns.
Defaults to "campaigns" which is what the API returns. It's expected to be
a list-column of data.frames}

\item{query_info}{The names of the columns that hold the query information. passed to \code{\link[=query_df_gen]{query_df_gen()}}}

\item{endpoint}{the name of the endpoint to query. passed to \code{\link[=query_df_gen]{query_df_gen()}}}

\item{request_col_name}{name of column holding the responses}
}
\value{
A data.frame just like the \code{starting_df}, but with a list-column
which contains the response.
}
\description{
Download the result of every query described in list-column
}
\details{
This function is designed to work with the same kind of object returned by
\code{\link[=download_sites_sf]{download_sites_sf()}}. However it does not need to be an sf data.frame. Any
data.frame with a column containing campaign information \emph{should} work. Very
often a table like this has a list-column of other, smaller dataframes. These
describe the info you need to send to the API to get a table related to the
first.
}
