% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sites_sf.R
\name{download_sites_sf}
\alias{download_sites_sf}
\title{Get rcoleo sites as sf dataframe}
\usage{
download_sites_sf(...)
}
\arguments{
\item{`...`}{Arguments to \code{\link[=get_sites]{get_sites()}}}
}
\description{
Download the site dataset and its metadata, and make it an sf dataframe.
}
\details{
The response from \code{\link[=get_sites]{get_sites()}} comes back as a list of simple data.frames.
However, for plotting, it is very convenient to work with \code{sf} data.frames.
This function facilitates the process by doing three things:
\itemize{
\item download all the sites
\item combine multiple pages in response, if any
\item pass the response through \code{\link[sf:st]{sf::st_point()}} and \code{\link[sf:st_as_sf]{sf::st_as_sf()}} to return a sf data.frame
}
}
