% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_format_utils.R
\name{format_spatial}
\alias{format_spatial}
\title{Formate les donnees avec leurs cordonnées spatiales.}
\usage{
format_spatial(df_to_inject)
}
\arguments{
\item{df_to_inject}{data.frame àinjecter}
}
\description{
La localisation des données doivent être formatées correctement pour être
injectées dans coleo.
}
\details{
Cette fonction est appelée par \code{\link[rcoleo]{coleo_prep_input_data}}.
}
