% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_all_obs.R
\name{get_all_obs}
\alias{get_all_obs}
\title{Retourne les observations présentes dans la base de données}
\usage{
get_all_obs()
}
\value{
returns a dataframe with all the observations, and the type of campaign that yielded them
}
\description{
Retourne les observations présentes dans la base de données
}
\examples{
\dontrun{
get_all_obs()
}
}
