% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_obs.R
\name{get_all_observations_from_a_site}
\alias{get_all_observations_from_a_site}
\title{Download all the observations of a specific site}
\usage{
get_all_observations_from_a_site(site_df, token = bearer())
}
\arguments{
\item{site_df}{a dataframe containing site information}

\item{token}{authentification information}
}
\description{
Starting from a site data.frame, request all the observations
}
\details{
this function assumes you have a site dataframe of the kind returned by
\code{\link[=get_sites]{get_sites()}} or \code{\link[=download_sites_sf]{download_sites_sf()}}. It is a simple manual wrapper around one workflow.
}
