% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_campaign_summary.R
\name{get_campaign_summary}
\alias{get_campaign_summary}
\title{get the campaign summary for a site}
\usage{
get_campaign_summary(campaign_type = NULL, site_code = NULL)
}
\arguments{
\item{campaign_type}{campaign type. Can be any valid campaign name, see \code{\link[=validate_campaign_type]{validate_campaign_type()}}}

\item{site_code}{any site code. Must be a single value}
}
\value{
data.frame of campaigns and years (for a site) or sites and years (for a campaign)
}
\description{
get the campaign summary for a site
}
\details{
This function gives a list of all sampling campaigns ever collected at a site.
You must supply either the type of campaign, the code of the site, or both.
}
