% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_cells.R
\name{get_cells}
\alias{get_cells}
\title{Obtenir les informations sur une cellule depuis l'API de coleo}
\usage{
get_cells(ids = ids, ...)
}
\arguments{
\item{ids}{est un vecteur contenant les identifiants des cellules que l'on désire obtenir. Si ids n'est pas spécifié, la fonction retournera l'ensemble des cellules présentes dans la base de données.}
}
\value{
Un objet \code{list}, dont chacun des niveaux corresponds à la réponse de l'API. La réponse peut être de classe \code{getError} ou \code{getSuccess}.
}
\description{
Obtenir les informations sur une cellule depuis l'API de coleo
}
\examples{
get_cells(ids=c('111_91'))
get_cells()
}
