% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_ids.R
\name{get_ids}
\alias{get_ids}
\title{Fonction generique permettant le retrait d'une entrée par son identifiant unique}
\usage{
get_ids(ids = NULL, endpoint = NULL, arg = "code", ...)
}
\arguments{
\item{ids}{est un vecteur contenant les identifiants uniques que l'on désire obtenir. Si ids n'est pas spécifié, la fonction retournera l'ensemble des sites présents dans la base de données.}

\item{endpoint}{Point d'entrée pour le retrait des données. Un point d'entrée peut être vu comme une table de la base de données.}

\item{...}{Arguments de la fonction generique \code{\link[httr:GET]{httr::GET()}}}
}
\value{
Retourne une objet de type \code{list} contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est \code{getSuccess} ou \code{getError}. Une réponse de classe \code{getSuccess} est une liste à deux niveaux composé du contenu (\code{body}), et la réponse \link[httr:response-class]{httr::response}. Une réponse de classe \code{getError}.
}
\description{
Fonction generique permettant le retrait d'une entrée par son identifiant unique
}
\details{
Les points d'accès de l'API sont énuméré dans l'environment de coléo, voir \code{print(rce$endpoints)}
}
