% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_lures.R
\name{get_lures}
\alias{get_lures}
\title{Retourne les entrées pour les leurres utilisées dans le cadre d'une campagne présents dans la base de données}
\usage{
get_lures(campaign_id = NULL, ...)
}
\arguments{
\item{campaign_id}{\code{integer} vecteur contenant les identifiants uniques de la campagne}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\value{
Retourne une objet de type \code{list} contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est \code{getSuccess} ou \code{getError}. Une réponse de classe \code{getSuccess} est une liste à deux niveaux composé du contenu (\code{body}), et la réponse \link[httr:response]{httr::response}. Une réponse de classe \code{getError} dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.
}
\description{
Retourne les entrées pour les leurres utilisées dans le cadre d'une campagne présents dans la base de données
}
\examples{
\dontrun{
get_lures()
} 
}
