% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_richness.R
\name{get_richness}
\alias{get_richness}
\title{Download}
\usage{
get_richness(
  campaign_type = NULL,
  site_type = NULL,
  site_code = NULL,
  by_site_type = NULL,
  by_campaign_type = NULL
)
}
\arguments{
\item{campaign_type}{any campaign type. see \code{\link[=validate_campaign_type]{validate_campaign_type()}}. Can be a partial match.}

\item{site_type}{any site type.}

\item{site_code}{a site_code.}

\item{by_site_type}{whether to summarize across site_types.}

\item{by_campaign_type}{whether to summarize across campaigns}
}
\value{
data.frame with richness values. Richness is simply calculated as unique species names that are not in category "autres".
}
\description{
get a datafram of unique site and species occurences for any type of sampling campaign.
}
