% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_samples.R
\name{get_samples}
\alias{get_samples}
\title{Retourne les entrées pour les échantillons présents dans la base de données}
\usage{
get_samples(sample_code = NULL, ...)
}
\arguments{
\item{sample_code}{\code{character} vecteur contenant les codes d'identification unique des échantillons}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\value{
Retourne une objet de type \code{list} contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est \code{getSuccess} ou \code{getError}. Une réponse de classe \code{getSuccess} est une liste à deux niveaux composé du contenu (\code{body}), et la réponse \link[httr:response]{httr::response}. Une réponse de classe \code{getError} dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.
}
\description{
Retourne les entrées pour les échantillons présents dans la base de données
}
\examples{
get_samples(trap_code=c('2016-0026'))
head(get_samples())
}
