% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_sites.R
\name{get_sites}
\alias{get_sites}
\title{Retourne les sites d'une cellule.}
\usage{
get_sites(query = NULL, ...)
}
\arguments{
\item{d}{une liste contenant pour chaque niveau l'information d'une cellule. L'informations qui peut être saisie sur une cellule est accessible à TODO http://XXX.}
}
\value{
Un objet de la classe \emph{responses} (liste) contenant les codes de status (201 ou 500). Le code 201 signifie que la ou les cellules ont bien été ajouté à la base de données alors que le code 500 siginifie qu'il y a eut une ou des erreurs.
}
\description{
Retourne les sites d'une cellule.
}
\examples{
get_sites()
}
