% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_species.R
\name{get_species}
\alias{get_species}
\title{Retourne les entrées d'espèces présentes dans la base de données}
\usage{
get_species(query = NULL)
}
\description{
Retourne les entrées d'espèces présentes dans la base de données
}
\examples{
get_sites()
}
