% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_site.R
\name{get_species_site}
\alias{get_species_site}
\title{Download one species by site matrix}
\usage{
get_species_site(campaign_type)
}
\arguments{
\item{campaign_type}{any campaign type. see \code{\link[=validate_campaign_type]{validate_campaign_type()}}. Can be a partial match.}
}
\value{
data. frame of species occurences in sites according to the database.
}
\description{
get a datafram of unique site and species occurences for any type of sampling campaign.
}
