% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_techs.R
\name{get_techs}
\alias{get_techs}
\title{Obtenir l'ensemble des techiciens attachés à une campagne}
\usage{
get_techs(name = NULL, lastname = NULL, ...)
}
\arguments{
\item{name}{\code{vector} contenant le prénom du technicien (sensible à la case)}

\item{lastname}{\code{vector} contenant le nom du technicien (sensible à la case)}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\description{
Obtenir l'ensemble des techiciens attachés à une campagne
}
\examples{
get_techs(name="Caroline")
}
