% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{is.coleoPostResp}
\alias{is.coleoPostResp}
\title{Test sur la classe \code{coleoPostResp} (Objet S3)}
\usage{
is.coleoPostResp(x)
}
\arguments{
\item{x}{Objet à tester}
}
\description{
Test sur la classe \code{coleoPostResp} (Objet S3)
}
