% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{is_ascii}
\alias{is_ascii}
\title{Validate the presence of non ASCII characters in given columns of a dataframe}
\usage{
is_ascii(data, dat_names)
}
\arguments{
\item{data}{A dataframe containing the data to be validated.}

\item{dat_names}{A vector of column names in the dataframe.}
}
\value{
A message if non ASCII characters are found.
}
\description{
Validate the presence of non ASCII characters in given columns of a dataframe
}
