% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{is_punctuation}
\alias{is_punctuation}
\title{Validate the presence of punctuation in given columns of a dataframe}
\usage{
is_punctuation(data, dat_names)
}
\arguments{
\item{data}{A dataframe containing the data to be validated.}

\item{dat_names}{A vector of column names in the dataframe.}
}
\value{
A message if punctuation is found.
}
\description{
This function checks for the presence of punctuation characters in specified columns of a dataframe. Accepts \code{|} as a separator for taxon names.
}
