% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{missing_obs}
\alias{missing_obs}
\title{Fonction pour calculer le nombre d'observations qui échoueront lors de l'injection}
\usage{
missing_obs(data, nvals)
}
\arguments{
\item{nvals}{Le nombre total de campagnes.}

\item{df_id}{Le dataframe contenant les données à injecter.}
}
\value{
Le message mis à jour.
}
\description{
Cette fonction calcule le nombre d'observations qui échoueront lors de l'injection
et met à jour le message en conséquence.
}
