% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{new_vegetation_transect_campaigns}
\alias{new_vegetation_transect_campaigns}
\title{Fonction pour calculer le nombre de nouvelles campagnes de transects de végétation}
\usage{
new_vegetation_transect_campaigns(data, nvals)
}
\arguments{
\item{nvals}{Le nombre total de campagnes.}

\item{df_id}{Le dataframe contenant les données à injecter.}
}
\value{
Le message mis à jour.
}
\description{
Cette fonction calcule le nombre de nouvelles campagnes et de campagnes existantes dans le dataframe,
et met à jour le message en conséquence.
}
