% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_attributes.R
\name{post_attributes}
\alias{post_attributes}
\title{Publication d'attribut(s) dans la base de données de Coléo}
\usage{
post_attributes(data, ...)
}
\arguments{
\item{data}{une liste ou chacun des niveaux corresponds aux données attribuées au ednpoint.}

\item{...}{httr options; arguments de la fonction \code{httr::POST()}}
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{attributes} de l'API de Coleo
}
