% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_cells.R
\name{post_cells}
\alias{post_cells}
\title{Publication d'une cellule sur la base de données de Coléo}
\usage{
post_cells(data)
}
\arguments{
\item{data}{une liste ou chacun des niveaux corresponds aux données attribuées au ednpoint.}
}
\value{
Un objet \code{list}, dont chacun des niveaux corresponds à la réponse de l'API. La réponse peut être de classe \code{postError} ou \code{postSuccess}.
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{cells} de l'API de Coleo
}
\seealso{
\code{\link{post_gen}} pour la structure de sortie de la fonction.
}
