% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_devices.R
\name{post_devices}
\alias{post_devices}
\title{Publication du matériel utilisé dans le cadre de la campagne  dans l'API de coléo}
\usage{
post_devices(data, ...)
}
\arguments{
\item{data}{une liste ou chacun des niveaux corresponds aux données attribuées au ednpoint.}

\item{...}{httr options; arguments de la fonction \code{httr::POST()}}
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{devices} de l'API de Coleo
}
