% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_gen.R
\name{post_gen}
\alias{post_gen}
\title{Fonction générique pour envoyer de l'information vers l'API de Coléo}
\usage{
post_gen(endpoint, singleton, ...)
}
\arguments{
\item{endpoint}{Point d'entrée pour le retrait des données.}

\item{singleton}{Données à injecter à l'intérieur de la base de données.}
}
\description{
Fonction générique pour envoyer de l'information vers l'API de Coléo
}
\examples{
data(cells)
post_gen("/cells",cells[1])
# Note: Cette fonction ne prend que des singletons
}
