% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_gen.R
\name{post_gen_list}
\alias{post_gen_list}
\title{Fonction générique pour envoyer de l'information vers l'API de Coléo}
\usage{
post_gen_list(data_list, endpoint, ...)
}
\arguments{
\item{data_list}{list contenant des donnees a injecter. voir \link{post_gen}}

\item{endpoint}{Point d'entrée pour le retrait des données.}

\item{...}{httr options; arguments de la fonction \code{httr::POST()}}
}
\value{
Retourne une objet de type \code{list} contenant les réponses de l'API.
}
\description{
Fonction générique pour envoyer de l'information vers l'API de Coléo
}
