% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_samples.R
\name{post_samples}
\alias{post_samples}
\title{Publication d'un échantillon dans l'API de coléo}
\usage{
post_samples(data, ...)
}
\arguments{
\item{data}{une liste ou chacun des niveaux corresponds aux données attribuées au ednpoint.}

\item{...}{httr options; arguments de la fonction \code{httr::POST()}}
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{samples} de l'API de Coleo
}
