% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_sites.R
\name{post_sites}
\alias{post_sites}
\title{Publication d'un site sur la base de données de Coléo}
\usage{
post_sites(data)
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{sites} de l'API de Coleo
}
