% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_traps.R
\name{post_traps}
\alias{post_traps}
\title{Publication d'un piège dans l'API de coléo}
\usage{
post_traps(data, ...)
}
\arguments{
\item{data}{une liste ou chacun des niveaux corresponds aux données attribuées au ednpoint.}

\item{...}{httr options; arguments de la fonction \code{httr::POST()}}
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{traps} de l'API de Coleo
}
