% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gen.R
\name{query_df_gen}
\alias{query_df_gen}
\title{Generate an API call from a data frame}
\usage{
query_df_gen(query_df, query_info, endpoint, ...)
}
\arguments{
\item{query_df}{the data.frame containing the query information. Usually from
a call to \code{\link[=get_gen]{get_gen()}}}

\item{query_info}{a character vector of column names in \code{query_df} which will
be passed to the API}

\item{endpoint}{the API endpoint to send all this to}

\item{...}{extra arguments to \code{\link[=get_gen]{get_gen()}}}
}
\value{
a data.frame containing the body of the responses.
}
\description{
Take a data.frame that stores query information and send a GET request with
that info
}
\details{
The coleo API returns a data.frame for most reponses. This data.frame can in
turn become the input for another request. This function facilitates that
process
}
