% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gen.R
\name{query_gen}
\alias{query_gen}
\title{query one endpoint and get a dataframe}
\usage{
query_gen(endpoint, ...)
}
\arguments{
\item{endpoint}{the name of the endpoint}

\item{...}{extra arguements to get_gen}
}
\value{
a data.frame of endpoint answers
}
\description{
query one endpoint and get a dataframe
}
