% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_get.R
\name{sf_cells}
\alias{sf_cells}
\title{Fonction retournant les cellules dans un objet de classe spatiale (sf)}
\usage{
sf_cells(cell_code = NULL, name = NULL, ...)
}
\arguments{
\item{cell_code}{\code{character} vecteur contenant les identifiants uniques que l'on désire obtenir. Si cell_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.}

\item{name}{\code{character} vecteur contenant les noms de la cellules que l'on veut obtenir.}

\item{...}{arguments de la fonction \code{httr::GET()}}
}
\description{
Fonction retournant les cellules dans un objet de classe spatiale (sf)
}
\seealso{
\code{\link[=get_cells]{get_cells()}}
}
