% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_get.R
\name{sf_landmarks}
\alias{sf_landmarks}
\title{Fonction retournant les repères utilisé dans le cadre d'une campagne dans un objet de classe spatiale (sf)}
\usage{
sf_landmarks(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)
}
\arguments{
\item{site_code}{\code{character} vecteur contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)}

\item{opened_at}{\code{character} vecteur contenant les dates de début de campagnes (ex. 2017-01-30)}

\item{closed_at}{\code{character} vecteur contenant les dates de fin de campagnes (ex. 2017-01-30)}

\item{type}{\code{character} vecteur contenant le type de campagnes d'inventaires réalisé (ex. végétation)}

\item{...}{arguments de la fonction \code{httr::GET()}}
}
\description{
Fonction retournant les repères utilisé dans le cadre d'une campagne dans un objet de classe spatiale (sf)
}
\seealso{
\code{\link[=get_landmarks]{get_landmarks()}}
}
