% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{sites_sf}
\alias{sites_sf}
\title{Fonction retournant les sites dans un objet de classe spatiale (sf)}
\usage{
sites_sf(...)
}
\arguments{
\item{...}{arguments de la fonction \code{httr::GET()} et \code{rcoleo::get_sites()}}
}
\description{
Fonction retournant les sites dans un objet de classe spatiale (sf)
}
\seealso{
\code{\link[=get_sites]{get_sites()}}
}
