% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_campaign_type.R
\name{validate_campaign_type}
\alias{validate_campaign_type}
\title{Validate user input of campaign type}
\usage{
validate_campaign_type(campaign_input)
}
\arguments{
\item{campaign_input}{Must be one of 'végétation', 'végétation_transect',
'sol', 'acoustique', 'phénologie', 'mammifères', 'papilionidés', 'odonates',
'insectes_sol', 'ADNe', 'zooplancton', 'sol', 'décomposition_sol',
'température_eau', 'température_sol', 'marais_profondeur_température'.
NOTE that partial matching is enabled: you can input "v" and get
"végétation" or "pap" and get "papilionidés"}
}
\value{
the full name of the campaign type, suitable for making another api
request.
}
\description{
Matches input against the valid campaign types. These are stored in an
internal table at \code{coleo_get_name_table.R}. If a new campaign type is added, this
will have to change.
}
