% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_campaign_type.R
\name{validate_site_type}
\alias{validate_site_type}
\title{Validate user input of site type}
\usage{
validate_site_type(site_input)
}
\arguments{
\item{site_input}{Must be one of "végétation", "papilionidés",
"acoustique", "insectes_sol", "mammifères", "odonates" or "zooplancton".
NOTE that partial matching is enabled: you can input "v" and get
"végétation" or "pap" and get "papilionidés"}
}
\value{
the full name of the site type, suitable for making another api
request.
}
\description{
Matches input against the valid site types. These are stored in an
internal table at \code{coleo_get_name_table.R}. If a new site type is added, this
will have to change.
}
