% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{as.data.frame.coleoGetResp}
\alias{as.data.frame.coleoGetResp}
\title{Transformer la liste de réponses en data.frame}
\usage{
\method{as.data.frame}{coleoGetResp}(x, row.names, optional, ...)
}
\arguments{
\item{x}{liste de réponse des fonctions de GET/POST}

\item{row.names}{inherits from data.frame. unused}

\item{optional}{inherits from data.frame. unused}

\item{...}{arguments to as.data.frame}
}
\description{
Transformer la liste de réponses en data.frame
}
