% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{cells_sf}
\alias{cells_sf}
\title{Fonction retournant les cellules dans un objet de classe spatiale (sf)}
\usage{
cells_sf(...)
}
\arguments{
\item{...}{arguments de la fonction \code{httr::GET()} et \code{rcoleo::get_cells()}}
}
\description{
Fonction retournant les cellules dans un objet de classe spatiale (sf)
}
\seealso{
\code{\link[=get_cells]{get_cells()}}
}
