% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_cell.R
\name{coleo_cell_geom_fmt}
\alias{coleo_cell_geom_fmt}
\title{Convert cell from sf to list}
\usage{
coleo_cell_geom_fmt(cell_geom)
}
\arguments{
\item{cell_geom}{an sf object which describes one cell}
}
\value{
a nested list with the coordinates of the polygon
}
\description{
We need to inject a nested list structure into the database. When we read
data into R, we often do so with sf::read_sf(). This function converts the
latter to the former using geojsonio
}
