% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_format_utils.R
\name{coleo_format}
\alias{coleo_format}
\title{Formatage automatisé des colonnes d'un jeu de données à injecter dans coleo}
\usage{
coleo_format(dataFrame)
}
\arguments{
\item{dataFrame}{un jeu de données}
}
\description{
Formatage automatisé des colonnes d'un jeu de données à injecter dans coleo
}
