% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_format_utils.R
\name{coleo_format}
\alias{coleo_format}
\title{Formatage automatisé des colonnes d'un jeu de données à injecter dans coleo}
\usage{
coleo_format(dataFrame)
}
\arguments{
\item{dataFrame}{un jeu de données}
}
\description{
Cette fonction formate les colonnes àla bonne classe. Elle est appelée par
les fonctions \code{\link[rcoleo]{coleo_read_template}} et \code{\link[rcoleo]{coleo_read_csv}}.
}
