% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_required_tables.R
\name{coleo_get_required_tables}
\alias{coleo_get_required_tables}
\title{Table des tables requises par type de campagne}
\usage{
coleo_get_required_tables()
}
\value{
a tbl_df showing which tables are required for which injection efforts.
}
\description{
Table des tables requises par type de campagne
}
