% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject}
\alias{coleo_inject}
\title{Injecte un jeu de données dans Coleo.}
\usage{
coleo_inject(df, media_path = NULL)
}
\arguments{
\item{df}{Un data.frame validé par \code{coleo_validate}}

\item{media_path}{NULL par défault. Requis lorsqu'il y a des fichiers médias
à injecter. Doit être le chemin local vers les fichiers médias à injecter.}
}
\value{
Un message de succès ou d'échec de l'injection des lignes du
data.frame par table et le data.frame retourné par \code{coleo_injection_perform}.
}
\description{
Principale fonction d'injection de données dans Coleo. Cette fonction
accepte un data.frame validé par \code{coleo_validate} et performe
automatiquement l'injection de toutes les tables dans Coleo.
}
