% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_adne_landmarks}
\alias{coleo_inject_adne_landmarks}
\title{Inject obs_edna table of ADNe campaigns into coleo}
\usage{
coleo_inject_adne_landmarks(df_id, failures)
}
\arguments{
\item{df_id}{a dataframe with campaign_id column}

\item{failures}{Boolean. Failures within the injection process}
}
\value{
a dataframe with landmarks ids
}
\description{
Takes a valid dataframe and performs autonomously the injection of obs_edna table
}
