% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_general}
\alias{coleo_inject_general}
\title{Fonction génàrale pour l'injection dans la base de donnàes Coleo via l'API.}
\usage{
coleo_inject_general(..., endpoint)
}
\arguments{
\item{...}{Donnàes à injecter. Peut contenir des valeurs NA ou NULL; elles seront retiràes avant l'injection.}

\item{endpoint}{Nom du endpoint de l'API où injecter.}
}
\value{
Un objet httr2 request, prêt à être envoyé à l'API Coleo.
}
\description{
Fonction génàrale pour l'injection dans la base de donnàes Coleo via l'API.
}
