% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_general}
\alias{coleo_inject_general}
\title{Fonction générale pour l'injection dans la base de données Coleo via l'API.}
\usage{
coleo_inject_general(endpoint, schema = "public", ...)
}
\arguments{
\item{endpoint}{Nom du endpoint de l'API où injecter.}

\item{schema}{Schema qui contient les tables de l'appel.}

\item{...}{Données à injecter. Peut contenir des valeurs NA ou NULL; elles seront retirées avant l'injection.}
}
\value{
Un objet httr2 request, prêt à être envoyé à l'API Coleo.
}
\description{
Fonction générale pour l'injection dans la base de données Coleo via l'API.
}
