% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_mam_lures}
\alias{coleo_inject_mam_lures}
\title{Inject lures table of mammifère campaigns into coleo}
\usage{
coleo_inject_mam_lures(df_id, failures)
}
\arguments{
\item{df_id}{a dataframe with campaign_id column}

\item{failures}{Boolean. Failures within the injection process}
}
\value{
a dataframe with lure ids
}
\description{
Takes a valid dataframe and performs autonomously the injection of lures table
}
