% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_media}
\alias{coleo_inject_media}
\title{Injection of media files into the Coleo server}
\usage{
coleo_inject_media(df_id, server_dir = "observation", file_dir)
}
\arguments{
\item{df_id}{A dataframe with *_id column(s)}

\item{dir}{The name of the directory you are injecting (campaign, observation, site).}

\item{file_path}{The local path to the media file to inject.}
}
\value{
an httr2 request, ready to be performed.
}
\description{
Injection of media files into the Coleo server
}
