% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_ref_species}
\alias{coleo_inject_ref_species}
\title{Injection des colonnes taxa_names dans la table ref_species de coleo.}
\usage{
coleo_inject_ref_species(taxa_col, schema = "public")
}
\arguments{
\item{taxa_col}{Un vecteur de taxa_names à injecter.}

\item{schema}{Schéma sur lequel faire la requête. Par défaut, le schéma
public est utilisé.}
}
\description{
Accepte un vecteur de noms de taxons et performe l'injection de la table
ref_species.
}
\details{
Cette fonction est silencieuse et ne retourne aucun message. Tous les taxons
seront injectés dans la table ref_species, mais ceux déjà présent
retourneront une erreur.

Cette fonction est utilisée par \code{\link[rcoleo]{coleo_inject}} et
\code{\link[rcoleo]{coleo_inject_mam_landmarks}}.
}
