% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_ref_species}
\alias{coleo_inject_ref_species}
\title{Injection des colonnes taxa_names dans la table ref_species de coleo.}
\usage{
coleo_inject_ref_species(taxa_col)
}
\arguments{
\item{taxa_col}{Un vecteur de taxa_names à injecter.}
}
\description{
Accepte un vecteur de noms de taxons et performe l'injection de la table
ref_species.
}
\details{
Cette fonction est silencieuse et ne retourne aucun message. Tous les taxons
seront injectés dans la table ref_species, mais ceux déjà présent
retourneront une erreur.

Cette fonction est utilisée par \code{coleo_inject} et
\code{coleo_inject_mam_landmarks}.
}
