% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_inject_table}
\alias{coleo_inject_table}
\title{Inject tables into coleo}
\usage{
coleo_inject_table(df_id, failures, table)
}
\arguments{
\item{df_id}{a dataframe with *_id column(s)}

\item{failures}{Boolean. Failures within the injection process}

\item{table}{a coleo table name}
}
\value{
a dataframe with lure ids
}
\description{
Takes a valid dataframe and performs autonomously the injection of data into appropriate table
}
