% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_inject.R
\name{coleo_injection_final}
\alias{coleo_injection_final}
\title{Finalize coleo injection}
\usage{
coleo_injection_final(df)
}
\arguments{
\item{df}{dataframe produced by \code{coleo_injection_execute}}
}
\value{
a data.frame. All the injected columns are gone, and it has no groups
or nested data. the ID of the records just injected is stored in a column
with the correct name
}
\description{
After successful injection, process the dataframe: extract the new ID from
the injected records, store it in a new column, and drop all the injected
columns
}
