% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_prep_input_data.R
\name{coleo_prep_input_data}
\alias{coleo_prep_input_data}
\title{Preparer les données pour injection}
\usage{
coleo_prep_input_data(df, db_table)
}
\arguments{
\item{df}{le dataframe à preparer}

\item{db_table}{la table ciblée dans la base de données}
}
\value{
un tbl, nested
}
\description{
Preparer les données pour injection
}
