% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_read.R
\name{coleo_read_csv}
\alias{coleo_read_csv}
\title{Read csv data and format columns to coleo requirements}
\usage{
coleo_read_csv(fileName)
}
\arguments{
\item{fileName}{Local path to csv file.}
}
\value{
a data.frame with formated columns
}
\description{
Read csv data and format columns to coleo requirements
}
