% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_read.R
\name{coleo_read_template}
\alias{coleo_read_template}
\title{Importe les données à injecter depuis un template coleo}
\usage{
coleo_read_template(templatePath)
}
\arguments{
\item{templatePath}{Chemin local vers le document csv.}
}
\value{
Un data.frame avec les colonnes formatées pour l'injection à coleo.
}
\description{
Importe les données à injecter depuis un template coleo
}
