% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_request_by_code}
\alias{coleo_request_by_code}
\title{Requête sur coléo par code}
\usage{
coleo_request_by_code(human_code, table, schema = "public", perform = TRUE)
}
\arguments{
\item{human_code}{Code de site ou de cellule.}

\item{table}{Table de la base de données à accéder. Seuls "cells" et "sites" sont supportés pour l'instant.}

\item{schema}{Schéma de la base de données à accéder. Par défaut, "public".}

\item{perform}{TRUE par default. Ne performe pas la requête et retourne l'object httr2 request si FALSE.}
}
\value{
si perform = TRUE, la réponse est retournée. Si perform = FALSE, la requête httr2 est retournée.
}
\description{
Plusieurs entrées de la base de données sont des codes qui sont lisibles par
les humains. Les codes de site suivent le format "CCC_CC_LCC"
(C = chiffre, L=lettre) et les codes de cellules suivent le format "CCC_CC".
Cette fonction reçoit un de ces codes et retourne les informations de cette
entrée.
}
