% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_request_general}
\alias{coleo_request_general}
\title{Requête générale sur coleo de type 'GET'}
\usage{
coleo_request_general(
  endpoint,
  perform = TRUE,
  response_as_df = FALSE,
  schema = "api",
  ...
)
}
\arguments{
\item{endpoint}{Nom du endpoint de l'API de coleo sur lequel la requête doit être effectuée}

\item{perform}{TRUE par default. Retourne un objet httr2 request et pas de requête effectuée si FALSE.}

\item{response_as_df}{FALSE par défaut. Retroune un data.frame si TRUE.}

\item{schema}{Schema qui contient les fonctions ou tables de l'appel}

\item{...}{Paramètres de requête pour la base de données coleo (dans leformat 'nom = valeur')}
}
\value{
Object httr2 response si perform = TRUE et un tibble si response_as_df = TRUE, un objet httr2 request si perform = FALSE.
}
\description{
Des paramètres rendu disponibles via l'API peuvent être 'count', ou 'offset'.
}
