% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_request_general}
\alias{coleo_request_general}
\title{General request for the coleo database}
\usage{
coleo_request_general(..., endpoint, perform = TRUE)
}
\arguments{
\item{...}{query parameters for the coleo database (in the format \code{name = value})}

\item{perform}{Should the request be performed? defaults to TRUE.}

\item{table}{name of database table to query}
}
\value{
httr2 response object if perform = TRUE, a request object if perform = FALSE.
}
\description{
General request for the coleo database
}
