% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_request.R
\name{coleo_resp_df}
\alias{coleo_resp_df}
\title{Converti un objet httr2 response en un tibble.}
\usage{
coleo_resp_df(resp, written_req, limit)
}
\arguments{
\item{resp}{objet httr2 response de coleo.}

\item{written_req}{Requête passée à l'API de coleo.}

\item{perform}{TRUE par défault. Ne performe pas la requête si FALSE.}
}
\value{
si perform = TRUE, la réponse est retournée. Si perform = FALSE, la requête httr2 est retournée.
}
\description{
Si la réponse contient plusieurs pages, la fonction accède à chacune et les
joints en un seul data.frame
}
