% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_return_cols.R
\name{coleo_return_cols}
\alias{coleo_return_cols}
\title{Trouver les colonnes, leur classe et les valeurs admissibles pour un type de campagne donné}
\usage{
coleo_return_cols(campaign_type, required.columns = FALSE)
}
\arguments{
\item{campaign_type}{un type de campagne}

\item{required.columns}{FALSE. Si TRUE, retourne seulement les colonnes requises pour procéder à l'injection}
}
\description{
Trouver les colonnes, leur classe et les valeurs admissibles pour un type de campagne donné
}
