% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_name_table.R
\name{coleo_return_required_tables}
\alias{coleo_return_required_tables}
\title{Trouver les tables requises pour un type de capagne donné}
\usage{
coleo_return_required_tables(camp_type)
}
\arguments{
\item{camp_type}{un type de campagne valide.}
}
\value{

}
\description{
Trouver les tables requises pour un type de capagne donné
}
