% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_get_name_table.R
\name{coleo_return_valid_campaigns}
\alias{coleo_return_valid_campaigns}
\title{Table des tables nécessaires pour chaque type de campagne}
\usage{
coleo_return_valid_campaigns()
}
\value{

}
\description{
Table des tables nécessaires pour chaque type de campagne
}
