% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_name_table_functions.R
\name{coleo_return_valid_site_types}
\alias{coleo_return_valid_site_types}
\title{Retourne un vecteur contenant les noms valides de type de site}
\usage{
coleo_return_valid_site_types()
}
\value{

}
\description{
Retourne un vecteur contenant les noms valides de type de site
}
