% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{coleo_validate}
\alias{coleo_validate}
\title{Valide la forme du jeu de données à injecter}
\usage{
coleo_validate(data)
}
\arguments{
\item{data}{jeu de données à valider}
}
\value{

}
\description{
Valide la forme du jeu de données à injecter
}
