% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleo_validate.R
\name{coleo_validate}
\alias{coleo_validate}
\title{Valide la forme du jeu de données à injecter}
\usage{
coleo_validate(data, media_path = NULL)
}
\arguments{
\item{data}{jeu de données à valider}

\item{media_path}{chemin vers le dossier contenant les médias à injecter}
}
\value{
Un message sur la validité du jeu de données.
}
\description{
Valide la forme du jeu de données à injecter
}
