% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spatial.R
\name{format_spatial}
\alias{format_spatial}
\title{Format les donnees avec leurs cordonnées spatiaux}
\usage{
format_spatial(df_to_inject)
}
\arguments{
\item{df_to_inject}{data.frame to be injected}
}
\description{
the location of data needs to be correctly formatted for injection into coleo
}
