% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_campaigns.R
\name{get_campaigns}
\alias{get_campaigns}
\title{Retourne les campagnes d'un site ou un ensemble de sites.}
\usage{
get_campaigns(id = NULL)
}
\arguments{
\item{id}{Un vecteur contenant la liste des codes de sites.}
}
\value{
Un objet de la classe \emph{data.frame} (tableau) contenant la liste des campagnes. Si le vecteur \emph{id} est d'une dimension supérieur à 1, une liste sera retourné. Chaque niveau de la liste correspondra à un site en particulier avec à l'intérieur un tableau decrivant les campagnes attachées à ce site.
}
\description{
Retourne les campagnes d'un site ou un ensemble de sites.
}
\examples{
get_campaigns()
}
