% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_campaigns.R
\name{get_campaigns}
\alias{get_campaigns}
\title{Obtenir l'ensemble des campagnes attachées à un site}
\usage{
get_campaigns(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)
}
\arguments{
\item{site_code}{\code{character} vecteur contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)}

\item{opened_at}{\code{character} vecteur contenant les dates de début de campagnes (ex. 2017-01-30)}

\item{closed_at}{\code{character} vecteur contenant les dates de fin de campagnes (ex. 2017-01-30)}

\item{type}{\code{character} vecteur contenant le type de campagnes d'inventaires réalisé (ex. végétation)}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\value{
Retourne une objet de type \code{list} contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est \code{getSuccess} ou \code{getError}. Une réponse de classe \code{getSuccess} est une liste à deux niveaux composé du contenu (\code{body}), et la réponse \link[httr:response]{httr::response}. Une réponse de classe \code{getError} dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.
}
\description{
Obtenir l'ensemble des campagnes attachées à un site
}
\examples{
\dontrun{
get_campaigns()
get_campaigns(type="végétation")
} 
}
