% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_campaigns.R
\name{get_campaigns}
\alias{get_campaigns}
\title{Obtenir l'ensemble des campagnes attachées à un site}
\usage{
get_campaigns(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)
}
\arguments{
\item{site_code}{\code{vector} contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)}

\item{opened_at}{\code{vector} contenant les dates de début de campagnes (ex. 2017-01-30)}

\item{closed_at}{\code{vector} contenant les dates de fin de campagnes (ex. 2017-01-30)}

\item{type}{\code{character} contenant le type de campagnes d'inventaires réalisé (ex. végétation)}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\description{
Obtenir l'ensemble des campagnes attachées à un site
}
\examples{
get_campaigns()
}
