% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_cells.R
\name{get_cells}
\alias{get_cells}
\title{Obtenir les informations sur les cellules depuis l'API de coleo}
\usage{
get_cells(cell_code = NULL, ...)
}
\arguments{
\item{cell_code}{est un vecteur contenant les identifiants uniques que l'on désire obtenir. Si cell_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\description{
Obtenir les informations sur les cellules depuis l'API de coleo
}
\examples{
get_cells(cell_code=c('111_91'))
get_cells()
}
