% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_cells.R
\name{get_cells}
\alias{get_cells}
\title{Retourne les campagnes d'un site ou un ensemble de sites.}
\usage{
get_cells(ids = NULL)
}
\arguments{
\item{d}{une liste contenant pour chaque niveau l'information d'une cellule. L'informations qui peut être saisie sur une cellule est accessible à TODO http://XXX.}
}
\value{
Un objet de la classe \emph{responses} (liste) contenant les codes de status (201 ou 500). Le code 201 signifie que la ou les cellules ont bien été ajouté à la base de données alors que le code 500 siginifie qu'il y a eut une ou des erreurs.
}
\description{
Retourne les campagnes d'un site ou un ensemble de sites.
}
\examples{
get_cells(codeName="110_112")
get_cells()
}
