% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_gen.R
\name{get_gen}
\alias{get_gen}
\title{Fonction générique pour retirer de l'information depuis l'API de Coléo}
\usage{
get_gen(endpoint = NULL, query = NULL, ...)
}
\arguments{
\item{endpoint}{\code{character} désignant le point d'entrée pour le retrait des données. Un point d'entrée peut être vu comme une table de la base de données.}

\item{query}{\code{list} de paramètres à passer avec l'appel sur le endpoint.}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\value{
Retourne une objet de type \code{list} contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est \code{getSuccess} ou \code{getError}. Une réponse de classe \code{getSuccess} est une liste à deux niveaux composé du contenu (\code{body}), et la réponse \link[httr:response]{httr::response}. Une réponse de classe \code{getError} dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.
}
\description{
Fonction générique pour retirer de l'information depuis l'API de Coléo
}
\details{
Les points d'accès de l'API sont énuméré dans l'environment de coléo, voir \code{print(rce$endpoints)}
}
\examples{
resp <- get_gen(rce$endpoints$cells)
length(resp) # Nombre de pages retourné par l'appel sur le point d'accès de l'API.
str(resp[[1]])
class(resp[[1]])
}
