% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_gen.R
\name{get_gen}
\alias{get_gen}
\title{Fonction générique pour retirer de l'information depuis l'API de Coléo}
\usage{
get_gen(endpoint, ...)
}
\arguments{
\item{endpoint}{Point d'entrée pour le retrait des données. Un point d'entrée peut être vu comme une table de la base de données}
}
\description{
Fonction générique pour retirer de l'information depuis l'API de Coléo
}
\examples{
get_gen(endpoints$cells)
}
