% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_sites.R
\name{get_sites}
\alias{get_sites}
\title{Obtenir les informations sur les sites depuis l'API de coleo}
\usage{
get_sites(site_code = NULL, ...)
}
\arguments{
\item{site_code}{est un vecteur contenant les identifiants uniques que l'on désire obtenir. Si site_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\description{
Obtenir les informations sur les sites depuis l'API de coleo
}
\examples{
get_sites(site_code=c('135_104_H01'))
get_sites()
}
