% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_species.R
\name{get_species}
\alias{get_species}
\title{Retourne les entrées d'espèces présentes dans la base de données}
\usage{
get_species(name = NULL, rank = NULL, vernacular_fr = NULL, ...)
}
\arguments{
\item{name}{\code{character} vecteur contenant les noms taxonomiques recherché (sensible à la case) - Obsolète si l'argument query est spécifié}

\item{rank}{\code{character} vecteur contenant les rangs taxonomiques recherché (sensible à la case) - Obsolète si l'argument query est spécifié}

\item{vernacular_fr}{\code{character} vecteur contenant les noms vernaculaires en francais des espèces recherchées (sensible à la case) - Obsolète si l'argument query est spécifié}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\value{
Retourne une objet de type \code{list} contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est \code{getSuccess} ou \code{getError}. Une réponse de classe \code{getSuccess} est une liste à deux niveaux composé du contenu (\code{body}), et la réponse \link[httr:response]{httr::response}. Une réponse de classe \code{getError} dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.
}
\description{
Retourne les entrées d'espèces présentes dans la base de données
}
\examples{
\dontrun{
get_species(vernacular_fr=c('Érable'))
head(get_species())
} 
}
