% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_by_site.R
\name{get_species_list}
\alias{get_species_list}
\title{get the species list for a site}
\usage{
get_species_list(campaign_type = "acoustique", site_code)
}
\arguments{
\item{campaign_type}{campaign type. Can be any valid campaign name, see \code{\link[=validate_campaign_type]{validate_campaign_type()}}}

\item{site_code}{any site code. can also be a vector}
}
\value{
the species list for a site, as a list
}
\description{
get the species list for a site
}
