% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_medias.R
\name{list_medias}
\alias{list_medias}
\title{Téléverse l'ensemble des médias appartenant à une campagnes}
\usage{
list_medias(x, ...)
}
\arguments{
\item{x}{vector of campaign ids}

\item{...}{httr options; arguments de la fonction \code{httr::GET()}}
}
\value{
Sauvegarde les images/sons attachés aux campagnes à une emplacement défini par l'utilisateur
}
\description{
Téléverse l'ensemble des médias appartenant à une campagnes
}
\examples{
\dontrun{
list_medias(x = c(101,104))
#Par exemple, on veut obtenir toutes les photos prises dans le cadre d'une campagne sur
# le site 141_108_F01 et portant sur les mammifères
get_medias(list_medias(x = c(101,104)))
}
}
