% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_cells.R
\name{post_cells}
\alias{post_cells}
\title{Retourne les campagnes d'un site ou un ensemble de sites.}
\usage{
post_cells(data)
}
\arguments{
\item{data}{une liste contenant pour chaque niveau l'information d'une cellule.}
}
\value{
Un objet de la classe \emph{responses} (liste) contenant les codes de status (201 ou 500).
}
\description{
Retourne les campagnes d'un site ou un ensemble de sites.
}
\examples{
data(cells)
post_cells(cells[1:10])
}
