% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_efforts.R
\name{post_efforts}
\alias{post_efforts}
\title{Publication d'un landmark dans l'API de coléo}
\usage{
post_efforts(data)
}
\arguments{
\item{data}{une liste ou chacun des niveaux corresponds aux données attribuées au ednpoint.}
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{efforts} de l'API de Coleo
}
