% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_sites.R
\name{post_sites}
\alias{post_sites}
\title{Publication d'un site sur la base de données de Coléo}
\usage{
post_sites(data)
}
\arguments{
\item{data}{une liste ou chacun des niveaux corresponds aux données attribuées à un site.}
}
\value{
Un objet \code{list}, dont chacun des niveaux corresponds à la réponse de l'API. La réponse peut être de classe \code{postError} ou \code{postSuccess}.
}
\description{
Cette fonction applique la méthode POST sur le point d'entrées \code{sites} de l'API de Coleo
}
\examples{
post_sites(list(list(cell_id = 1,code_site = "111_140_H01", date_open = "2017-08-31", type="forestier", off_station_code_id=NA)))
}
\seealso{
\code{\link{post_gen}} pour la structure de sortie de la fonction.
}
