% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{sites_sf}
\alias{sites_sf}
\title{Fonction retournant les sites dans un objet de classe spatiale (sf)}
\usage{
sites_sf(site_code = NULL, ...)
}
\arguments{
\item{site_code}{\code{character} vecteur contenant les identifiants uniques que l'on désire obtenir. Si site_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.}

\item{...}{arguments de la fonction \code{httr::GET()}}
}
\description{
Fonction retournant les sites dans un objet de classe spatiale (sf)
}
\seealso{
\code{\link[=get_sites]{get_sites()}}
}
